+++ src/Core/Mailable.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Niang\Core;
+
+abstract class Mailable
+{
+    abstract public function subject(): string;
+
+    abstract public function body(): string;
+}